chore: replace controller with service layer (#14)
* Add `findAllUsers`, `findUser`, `removeUser` and `createUser` service methods to controller.
* Remove unused repository `RepositoryInterface` interface.
* Add service layer to user controller create user method.
* Fix `createUser` and `removeUser` service methods.
* Update unit and integration tests after controller and service updates.

+++ src/Service/UserService.php
@@ -27,7 +27,12 @@ class UserService implements UserServiceInterface {
         return $this->userRepository->findAll();
     }
 
-    public function removeUser(User $user): void {
+    public function removeUser(int $userId): void {
+        $user = $this->userRepository->find($userId);
+        if(empty($user)) {
+            throw new UserNotFoundException($userId);
+        }
+
         $this->userRepository->remove($user, true);
     }
 
@@ -39,7 +44,7 @@ class UserService implements UserServiceInterface {
             $userCreate->getPhone(),
         );
 
-        return $this->userRepository->save($user);
+        return $this->userRepository->save($user, true);
     }
 
     public function updateUser(int $userId, UserCreate $userCreate): User {
@@ -57,7 +62,7 @@ class UserService implements UserServiceInterface {
             $user->setPassword($userCreate->getPassword());
         }
 
-        $this->userRepository->save($user);
+        $this->userRepository->save($user, true);
 
         return $user;
     }
